+++ platform/api/equinixmetal/api.go
@@ -102,6 +102,8 @@ type Options struct {
 	// Google Storage base URL for temporary uploads
 	// e.g. gs://users.developer.core-os.net/bovik/mantle
 	StorageURL string
+	// Metro is where you want your server to live.
+	Metro string
 }
 
 type API struct {
@@ -221,7 +223,7 @@ func (a *API) CreateOrUpdateDevice(hostname string, conf *conf.Conf, console Con
 	plog.Debugf("Created device: %q", deviceID)
 
 	if console != nil {
-		err := a.startConsole(deviceID, console)
+		err := a.startConsole(deviceID, device.Facility.Code, console)
 		consoleStarted = true
 		if err != nil {
 			a.DeleteDevice(deviceID)
@@ -522,6 +524,12 @@ func (a *API) createDevice(hostname, ipxeScriptURL, id string) (*packngo.Device,
 			plog.Infof("device rebooted: %s", id)
 		} else {
 			plog.Infof("Recycling is not possible, creating a new instance")
+			// if the Metro is set, we set the Facility to empty string in order
+			// to not conflict with Metro value.
+			if a.opts.Metro != "" {
+				a.opts.Facility = ""
+			}
+
 			device, response, err = a.c.Devices.Create(&packngo.DeviceCreateRequest{
 				ProjectID:     a.opts.Project,
 				Facility:      []string{a.opts.Facility},
@@ -532,6 +540,7 @@ func (a *API) createDevice(hostname, ipxeScriptURL, id string) (*packngo.Device,
 				IPXEScriptURL: ipxeScriptURL,
 				Tags:          []string{"mantle"},
 				AlwaysPXE:     alwaysPXE,
+				Metro:         a.opts.Metro,
 			})
 		}
 
@@ -551,13 +560,13 @@ func (a *API) createDevice(hostname, ipxeScriptURL, id string) (*packngo.Device,
 	return nil, fmt.Errorf("reached maximum number of retries to create/update a device: %w", err)
 }
 
-func (a *API) startConsole(deviceID string, console Console) error {
+func (a *API) startConsole(deviceID, facility string, console Console) error {
 	ready := make(chan error)
 
 	runner := func() error {
 		defer console.Close()
 
-		client, err := console.SSHClient("sos."+a.opts.Facility+".platformequinix.com", deviceID)
+		client, err := console.SSHClient("sos."+facility+".platformequinix.com", deviceID)
 		if err != nil {
 			return fmt.Errorf("couldn't create SSH client for %s console: %v", deviceID, err)
 		}